refactor, update: verstion update, naming refactor
