--- .idea/shelf/Uncommitted_changes_before_Checkout_at_2021-11-09__9_47__Default_Changelist_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_2021-11-09_오전_9_47_[Default_Changelist]" date="1636418937971" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2021-11-09_오전_9_47_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2021-11-09 오전 9:47 [Default Changelist]" />
-</changelist>
