--- .idea/shelf/Uncommitted_changes_before_Update_at_2021-11-07__1_08__Default_Changelist_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Update_at_2021-11-07_오후_1_08_[Default_Changelist]" date="1636258133171" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_2021-11-07_오후_1_08_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Update at 2021-11-07 오후 1:08 [Default Changelist]" />
-</changelist>
