--- .idea/shelf/Uncommitted_changes_before_Update_at_2021-11-07_오후_1_08_[Default_Changelist]/shelved.patch
@@ -1,100 +1,0 @@
-Index: .idea/gradle.xml
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
-<+><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<project version=\"4\">\r\n  <component name=\"GradleMigrationSettings\" migrationVersion=\"1\" />\r\n  <component name=\"GradleSettings\">\r\n    <option name=\"linkedExternalProjectsSettings\">\r\n      <GradleProjectSettings>\r\n        <option name=\"testRunner\" value=\"PLATFORM\" />\r\n        <option name=\"distributionType\" value=\"DEFAULT_WRAPPED\" />\r\n        <option name=\"externalProjectPath\" value=\"$PROJECT_DIR$\" />\r\n        <option name=\"modules\">\r\n          <set>\r\n            <option value=\"$PROJECT_DIR$\" />\r\n            <option value=\"$PROJECT_DIR$/app\" />\r\n          </set>\r\n        </option>\r\n        <option name=\"resolveModulePerSourceSet\" value=\"false\" />\r\n      </GradleProjectSettings>\r\n    </option>\r\n  </component>\r\n</project>
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
-diff --git a/.idea/gradle.xml b/.idea/gradle.xml
---- a/.idea/gradle.xml	(revision d7cbf457915cf44a70ea4c0f5a3e385db5a3b437)
-+++ b/.idea/gradle.xml	(date 1636256819036)
-@@ -4,7 +4,7 @@
-   <component name="GradleSettings">
-     <option name="linkedExternalProjectsSettings">
-       <GradleProjectSettings>
--        <option name="testRunner" value="PLATFORM" />
-+        <option name="testRunner" value="GRADLE" />
-         <option name="distributionType" value="DEFAULT_WRAPPED" />
-         <option name="externalProjectPath" value="$PROJECT_DIR$" />
-         <option name="modules">
-Index: .idea/misc.xml
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
-<+><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<project version=\"4\">\r\n  <component name=\"ProjectRootManager\" version=\"2\" languageLevel=\"JDK_1_7\" project-jdk-name=\"1.8\" project-jdk-type=\"JavaSDK\">\r\n    <output url=\"file://$PROJECT_DIR$/build/classes\" />\r\n  </component>\r\n  <component name=\"ProjectType\">\r\n    <option name=\"id\" value=\"Android\" />\r\n  </component>\r\n</project>
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
-diff --git a/.idea/misc.xml b/.idea/misc.xml
---- a/.idea/misc.xml	(revision d7cbf457915cf44a70ea4c0f5a3e385db5a3b437)
-+++ b/.idea/misc.xml	(date 1636257931372)
-@@ -1,6 +1,16 @@
- <?xml version="1.0" encoding="UTF-8"?>
- <project version="4">
--  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_7" project-jdk-name="1.8" project-jdk-type="JavaSDK">
-+  <component name="DesignSurface">
-+    <option name="filePathToZoomLevelMap">
-+      <map>
-+        <entry key="..\:/Users/kimta/AndroidStudioProjects/cooperation/MindCareApp/app/src/main/res/layout/fragment_home.xml" value="0.25364583333333335" />
-+        <entry key="..\:/Users/kimta/AndroidStudioProjects/cooperation/MindCareApp/app/src/main/res/layout/welcome_slide1.xml" value="0.25364583333333335" />
-+        <entry key="..\:/Users/kimta/AndroidStudioProjects/cooperation/MindCareApp/app/src/main/res/layout/welcome_slide2.xml" value="0.25364583333333335" />
-+        <entry key="..\:/Users/kimta/AndroidStudioProjects/cooperation/MindCareApp/app/src/main/res/layout/welcome_slide3.xml" value="0.25364583333333335" />
-+      </map>
-+    </option>
-+  </component>
-+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
-     <output url="file://$PROJECT_DIR$/build/classes" />
-   </component>
-   <component name="ProjectType">
-Index: .idea/jarRepositories.xml
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
-diff --git a/.idea/jarRepositories.xml b/.idea/jarRepositories.xml
-new file mode 100644
---- /dev/null	(date 1636257023897)
-+++ b/.idea/jarRepositories.xml	(date 1636257023897)
-@@ -0,0 +1,25 @@
-+<?xml version="1.0" encoding="UTF-8"?>
-+<project version="4">
-+  <component name="RemoteRepositoriesConfiguration">
-+    <remote-repository>
-+      <option name="id" value="central" />
-+      <option name="name" value="Maven Central repository" />
-+      <option name="url" value="https://repo1.maven.org/maven2" />
-+    </remote-repository>
-+    <remote-repository>
-+      <option name="id" value="jboss.community" />
-+      <option name="name" value="JBoss Community repository" />
-+      <option name="url" value="https://repository.jboss.org/nexus/content/repositories/public/" />
-+    </remote-repository>
-+    <remote-repository>
-+      <option name="id" value="BintrayJCenter" />
-+      <option name="name" value="BintrayJCenter" />
-+      <option name="url" value="https://jcenter.bintray.com/" />
-+    </remote-repository>
-+    <remote-repository>
-+      <option name="id" value="Google" />
-+      <option name="name" value="Google" />
-+      <option name="url" value="https://dl.google.com/dl/android/maven2/" />
-+    </remote-repository>
-+  </component>
-+</project>
-\ No newline at end of file
-Index: .idea/compiler.xml
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
-diff --git a/.idea/compiler.xml b/.idea/compiler.xml
-new file mode 100644
---- /dev/null	(date 1636257008690)
-+++ b/.idea/compiler.xml	(date 1636257008690)
-@@ -0,0 +1,6 @@
-+<?xml version="1.0" encoding="UTF-8"?>
-+<project version="4">
-+  <component name="CompilerConfiguration">
-+    <bytecodeTargetLevel target="1.8" />
-+  </component>
-+</project>
-\ No newline at end of file
