+++ app/src/main/java/com/teamopendata/mindcareapp/MainActivity.java
@@ -40,7 +40,7 @@ public class MainActivity extends BaseActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_setting, R.id.navigation_notifications)
+                R.id.navigation_home, R.id.navigation_map, R.id.navigation_graph, R.id.navigation_records)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         //NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
