+++ app/src/main/java/com/teamopendata/mindcareapp/ui/graph/GraphViewModel.java
@@ -1,14 +1,14 @@
-package com.teamopendata.mindcareapp.ui.setting;
+package com.teamopendata.mindcareapp.ui.graph;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class SettingViewModel extends ViewModel {
+public class GraphViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public SettingViewModel() {
+    public GraphViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is setting fragment");
     }
