+++ app/src/main/java/com/teamopendata/mindcareapp/ui/map/MapListAdapter.java
@@ -1,4 +1,4 @@
-package com.teamopendata.mindcareapp.ui.dashboard;
+package com.teamopendata.mindcareapp.ui.map;
 
 import android.content.Context;
 import android.view.LayoutInflater;
