+++ app/src/main/java/com/teamopendata/mindcareapp/ui/map/MapViewModel.java
@@ -1,16 +1,16 @@
-package com.teamopendata.mindcareapp.ui.notifications;
+package com.teamopendata.mindcareapp.ui.map;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class MapViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public MapViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("Hello World");
+        mText.setValue("This is dashboard fragment");
     }
 
     public LiveData<String> getText() {
