+++ app/src/main/java/com/teamopendata/mindcareapp/ui/records/RecordsFragment.java
@@ -1,0 +1,23 @@
+package com.teamopendata.mindcareapp.ui.records;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+
+import com.teamopendata.mindcareapp.R;
+
+public class RecordsFragment extends Fragment {
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_records, container, false);
+    }
+
+
+
+
+}
