+++ app/src/main/res/layout/fragment_graph.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/default_background"
-    tools:context=".ui.notifications.NotificationsFragment">
+    tools:context=".ui.records.RecordsFragment">
 
     <LinearLayout
         android:layout_marginTop="20dp"
