+++ app/src/main/res/layout/fragment_map.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/default_background"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context=".ui.map.MapFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
