+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -3,21 +3,21 @@
 
     <item
         android:id="@+id/navigation_home"
-        android:icon="@drawable/fragmenticon_home"
+        android:icon="@drawable/icon_home"
         android:title="홈" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/fragmenticon_map"
+        android:id="@+id/navigation_map"
+        android:icon="@drawable/icon_map"
         android:title="위치" />
 
     <item
-        android:id="@+id/navigation_notifications"
+        android:id="@+id/navigation_records"
         android:icon="@drawable/icon_record"
         android:title="작성" />
     <item
-        android:id="@+id/navigation_setting"
-        android:icon="@drawable/fragmenticon_graph"
+        android:id="@+id/navigation_graph"
+        android:icon="@drawable/icon_graph"
         android:title="그래프" />
 
 </menu>
