+++ app/src/main/res/menu/menu_base.xml
@@ -4,6 +4,6 @@
     <item
         android:id="@+id/menu_setting"
         android:icon="@drawable/setting_icon"
-        android:title="@string/title_setting"
+        android:title="@string/title_graph"
         app:showAsAction="always" />
 </menu>
