+++ app/src/main/res/values/strings.xml
@@ -2,9 +2,9 @@
     <string name="app_name">MindCareApp</string>
     <string name="title_activity_main">MainActivity</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
-    <string name="title_setting">Setting</string>
+    <string name="title_map">Map</string>
+    <string name="title_records">Records</string>
+    <string name="title_graph">Graph</string>
 
     <string name="title_activity_welcome">Home Screen</string>
     <string name="next">NEXT</string>
